Refactor seastate common tuner utilities into tuner module
